Added hourly sales panel.

+++ SharedResources/Results/HourlySalesResult.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace SharedResources
 {
-    public readonly struct HourlySalesResult(int hourOfDay, int uniqueMovies, int uniqueTheaters, int ticketSales)
+    public readonly struct HourlySalesResult(int hourOfDay, int uniqueMovies, int uniqueTheaters, decimal ticketSales)
     {
         public int HourOfDay { get; } = hourOfDay;
         public int MovieCount { get; } = uniqueMovies;
